# highlight: Avoid making unnecessary allocation