@bep 2020-10-16
A tip when running benchmarks is to turn off any "turbo boost". If you're using a tool such as my `gobench` which runs before/after in a sequence, with turbo boost on the first run is often artificial better. Since the delta is negative here, I have no doubt there is a positive effect of the above, I just thought I should mention it. 

My lessons learned with Go benchmarks:

* Turn off turbo boos
* Also try to not have a lot of other CPU stuff going on on your computer