@bep 2020-10-16
Oh, and the comment above is only relevant for time/cpu, which is not relevant at all for your bench.