@moorereason 2020-10-16
Agreed, but I was only looking for mallocs this time around.